Add test for device-display association by descriptor
This change introduces this new InputTest unit test,
`Configure_DeviceDescriptorAndDisplayUniqueId_CorrectlyMatches`, to
verify that an InputDevice is correctly associated with a display
using its device descriptor and unique ID. This test also verifies that
the association is reset when the policy rule is removed.

The `Configure_UniqueId_CorrectlyMatches` test is renamed to
`Configure_DeviceLocationAndDisplayUniqueId_CorrectlyMatches` to better
differentiate between association types. This test was also updated to
verify that the association is reset when the policy rule is removed.

Fix: 341356351
Flag: TEST_ONLY
Test: atest inputflinger_tests
Change-Id: I31348bac7848b2b154e212222ba961ae39d5f1bb