Throw errors if the requested pixel format doesn't work with DPU readback.
Current behavior is that when the client requests screenshot with DPU
readback. The requested pixel format will be ignored and the system will
return the screenshot with whatever that format DPU readback supported.

From the API design perspective, we should respect the requested pixel
format and throw errors if the request is incompatiable with DPU
readback.

Bug: 440394169
Test: atest ScreenCaptureTest
Flag: EXEMPT minor bug fix.

Change-Id: Ie0f5ab72296f743365b124e501ac7ceb78a4b1f1