SF: Plumb DisplayConnectionType to Scheduler::Display
Currently, the Scheduler does not take whether the display is internal
or external into account for any of its operations. However, we want to
bias the pacesetter selection logic to the external display. To do this,
we need to plumb the display connection type at registration time.

Flag: EXEMPT PURE_REFACTOR
Test: existing SchedulerTest
Bug: 440153446
Change-Id: Icd36ecf09cc2a125d23130e689fb238da56077be