SF: Prefer external displays as pacesetters given equal refresh rates
Change Scheduler::selectPacesetterDisplayLocked() to prefer external
displays over internal displays, if the refresh rates are roughly
equivalent; even if the internal display is the current pacesetter. The
churn is preferrable to keeping the internal display as the pacesetter
as the external display is more likely to be the main display in terms
of use.

Flag: EXEMPT BUGFIX
Test: New SchedulerTest case
Test: Manually verified w/ dumpsys on comet w/ ext display @60Hz
Bug: 440153446
Change-Id: I3dc68cc4ff81e18ce802133908ef360b6c74da9d