Remove the 5 frame wait for surfaceflinger timings
Currently we wait 5 frames before asking surface flinger for any timing
information. This breaks the CTS tests for the new
VK_EXT_present_timing. This limit can be removed as it was added a long
time ago and is outdated.

Test: presubmits, CTS run and timing values checked manually
Flag: NONE we are unlikely to get any useful signals from dogfood
Bug: 440824659
Change-Id: I02a5f4d25f0d591106241d11a99a8c16527344b7