tracing: use "stall then drop" policy
The current "stall" policy aborts the SF process if the stalling reaches
a certain timeout, assuming that the timeout is reached because of a
deadlock. However, in some rare cases the stall timeout can be reached
when the system is under pressure and traced can't consume trace packets
fast enough.

This commit changes the policy to "stall then drop". If the stalling
times out, packets are simply dropped (no process abort) and winscope
will warn the user about the missing tracing data.

Fix: 434058521
Test: presubmit
Flag: EXEMPT minor fix
Change-Id: I7adae4e71cff1107b29f2001710b62d9d14886bc