TouchInputMapper: Remove `hasAssociatedDisplay` parameter configuration
The parameters are not recomputed upon display changes, so we cannot
make any display-dependent calculations during parameter computation.

The `hasAssociatedDisplay` parameter in `TouchInputMapper::Parameters`
is removed. The logic for determining whether a touch device is
associated with a display is now derived directly from the device type,
unique display ID, and the associated viewport provided by the device
context.

Touch navigation devices will use a non-display viewport if no display
association is found, as it is the only device type in
TouchInputMapper that does not require display association.

Bug: 416791918
Test: Presubmit
Flag: EXEMPT cleanup, bug fix
Change-Id: Ic80adf25e15b6fd0e3d6d53f4f8be4f158d6fda8