Detect cycles in parent layer hierarchies
SF crashes when there is a cyclical hierarchy because
the existing cycle check only looks for cycles in the
root layer's children. The parent hierarchy is then
traversed when checking for secure layers, leading to
the crash. Modify the layer cycle method to check for
cycles in the parent hierarchy as well.

Bug: b/432594483
Test: atest SurfaceFlinger_test
Test: atest ScreenCaptureTest#captureCyclicalHierarchyInParentsFails
Flag: EXEMPT BUGFIX
Change-Id: I1826f0fecaebb3f1d5c9096e2d1ee4922afafa1a