Move screen dimensions into a separate struct
Previously, some callers would send in valid values for the width and
height of the screen (because this parameter is required for device
types like touchscreen), while others just sent in zeros.

Instead, make it more explicit in this CL - crash if the struct is not
provided for device types that require it, and allow it to be nullopt
for the other device types.

Bug: 439992405
Test: atest InputTests
Flag: EXEMPT PURE_REFACTOR
Change-Id: I25ee8bf88aea41d388c7f9ba2ae8a51087f19e53